Modifications for app publishing and bug fixing
Created some assets which would aid in the production release of the app.

Additionally added changed around some code to allow matches to always exist even on moving pieces but the match group they belong to would not be deleted if one of the pieces in the group was still moving.  This helps prevent pieces from being left behind to due movement discrepancies.