Made some minor bug fixes
After careful review, it was identified that pieces not in pattern were breaking without being actually touched while a pattern directly above them was being interacted with by the player.  This issue arose because as the pattern above was breaking, a piece loaded in off screen which happened to land perfectly in position to make a row WHEN it reached the position it was meant to be in after floating upward.  I believe I resolved the issue by putting the AddNewPieces function into the LateUpdate method so that the pieces would finish checking for patterns on non moving pieces before they detected the newly instantiated pieces.  Sort of confusing.  Also changed an iteration function to continue loop on null, while also retaining combo status if any pieces on the board were moving.  A LogError statement became annoying so I turned it off as it's irrelevant.